pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.0.1"
description = "TODO aplication made with Python's FastAPI framework and Hexagonal Architecture"
authors = ["Giovanni Armane <giovanniarmane@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["email_validator", "requests"], version = "^0.52.0"}
uvicorn = "^0.11.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
factory_boy = "^2.12.0"
Faker = "^4.0.1"
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
flake8-broken-line = "^0.1.1"
flake8-bugbear = "^20.1.4"
flake8-builtins = "^1.4.2"
flake8-comprehensions = "^3.2.2"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^0.2.4"
flake8-pytest-style = "^0.2.0"
mypy = "^0.761"
pep8-naming = "^0.9.1"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
pytest-factoryboy = "^2.0.3"
pytest-mock = "^2.0.0"
pytest-sugar = "^0.9.2"
python-dotenv = "^0.12.0"
pytest-asyncio = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
